pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctx"
version = "0.1.0"
description = "Git for LLM context. Snapshot a dying session, brief the next LLM in 30 seconds."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "gautam sharma" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
    "anthropic>=0.39",
    "pyperclip>=1.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
ctx = "ctx.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ctx.distiller" = ["prompts/*.md"]
